snapp-main/requirements.txt
PyQt6==6.6.1
fpdf2==2.6.0
